Aborting only at the end of the item; bump to 20190106.02
The UOL forums appear to have a server-side cache for the successful responses. By iterating over everything first, we pre-heat the cache. On the item retry, it then gets retrieved successfully.
This unfortunately means we retrieve everything twice, but without a decent solution for filtering the error responses out of the WARCs, this is at least a viable workaround which should let us grab a bit more content.